2007-07-14 Andreas Bertheussen <andreasmarcel@gmail.com>
	Getting rid of compiler warnings when running with -Wall.

	* uploadfile.c: Explicit casting to int when using the off_t as int in printf() calls.
	* Casio9860.c: Added error message on allocation failure in fx_doConnVer(), and removed unused variables in fx_sendPositive() and fx_sendNegative().
	* xfer9860.c: Removed unused variable, and set sourceFileName to be initialized on declaration.
	* SConstruct: Added -Wall to the default ccflags.

+++ xfer9860/trunk/src/uploadfile.c
@@ -50,7 +50,7 @@ int uploadFile(char* sourceFileName, char* destFileName, int throttleSetting) {
 
 	struct stat sourceFileStatus;
 	stat(sourceFileName, &sourceFileStatus);
-	printf("[I] File size:  %i byte(s)\n", sourceFileStatus.st_size);
+	printf("[I] File size:  %i byte(s)\n", (int)sourceFileStatus.st_size);
 
 	printf("[>] Setting up USB connection.. ");
 	struct usb_dev_handle *usb_handle;
@@ -81,7 +81,7 @@ int uploadFile(char* sourceFileName, char* destFileName, int throttleSetting) {
 
 	int packetCount = ceil(sourceFileStatus.st_size/MAX_DATA_PAYLOAD) + 1;
 	printf("\n[>] Starting transfer of %s to fls0, %i b, %i packets..\n",
-	       destFileName, sourceFileStatus.st_size, packetCount);
+	       destFileName, (int)sourceFileStatus.st_size, packetCount);
 	char *fData = (char*)calloc(MAX_DATA_PAYLOAD, sizeof(char));
 	char *sData = (char*)calloc(MAX_DATA_PAYLOAD*2, sizeof(char));
 	char *buffer = (char*)calloc((MAX_DATA_PAYLOAD*2)+18, sizeof(char));	// work buffer
